fix ItadakiDataEdict build failure on Windows
This project always failed on Windows (or at least on my machine)
because the call to `File.renameTo(File)` always failed, preventing
the final dictionary file from being created. Fix this simply copying
and deleting when the `renameTo` method fails. This can still possibly
fail and of course takes longer, but is not as buggy. Note in STDERR
that this fallback is being used. This could probably be more easily
resolved with Java NIO, but we are supporting Java 1.5.

Also make changes to make this easier to debug in the future: fail the
ant build when `java` exits with an error, throw a separate exception
if an old index file could not be deleted, and note in STDERR if the
temp file could not be removed. Delete the temporary index file in
the ant `clean` target in case it's still around.

+++ Seashell/src/org/itadaki/seashell/edict/EdictDictionary.java
@@ -23,8 +23,11 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.io.RandomAccessFile;
 import java.nio.BufferUnderflowException;
 import java.nio.ByteBuffer;
@@ -362,15 +365,45 @@ public class EdictDictionary {
 		indexChannel.close();
 
 		File indexFile = new File (this.dictionaryFileName + ".iidx");
-		indexFile.delete();
+		if(indexFile.exists() && !indexFile.delete()){
+			throw new IOException ("Could not delete old index file");
+		}
 		if (!temporaryIndexFile.renameTo(indexFile)) {
-			throw new IOException ("Failed to move temporary index file to final position");
+			//renameTo often fails on Windows, so try copying and deleting instead
+			System.err.println("Renaming temporary index file failed; copying instead");
+			copyFile(temporaryIndexFile, indexFile);
+			if(!temporaryIndexFile.delete()){
+				System.err.println("Could not delete temporary index file (" + temporaryIndexFile + ")");
+			}
 		}
 
 		loadIndex();
 
 	}
-
+	
+	/**
+	 * Copy a file. This might work when {@link File#renameTo(File)} fails.
+	 * @param source File to copy from
+	 * @param dest File to copy to
+	 * @throws IOException if copying fails
+	 */
+	private static void copyFile(File source, File dest)
+			throws IOException {
+		InputStream input = null;
+		OutputStream output = null;
+		try {
+			input = new FileInputStream(source);
+			output = new FileOutputStream(dest);
+			byte[] buf = new byte[1024];
+			int bytesRead;
+			while ((bytesRead = input.read(buf)) > 0) {
+				output.write(buf, 0, bytesRead);
+			}
+		} finally {
+			input.close();
+			output.close();
+		}
+	}
 
 	/**
 	 * Loads an existing index file
